requirements.txt
python-dotenv
tk
google-generativeai

